// DogdexAI/src/lib/types.ts
export interface BreedInfo {
  breed?: string
  group?: string
  slug?: string
  height?: string
  weight?: string
  life_expectancy?: string
  energy_level?: number
  trainability?: number
  shedding_level?: number
  temperament?: string[]
  description?: string
}